Automating Data Preparation
The goal of this script is to quickly read in raw data from a file, give the user an overview of the contents within, and then convert each column to have the correct datatype (i.e from "$1,234.55" to 1234.55) with minimal manual involvement. This is most suitable for smaller size datasets (< 1million rows).